streamlit/requirements.txt
streamlit==1.31.0
google-cloud-bigquery==3.14.1
plotly==5.18.0
pandas==2.1.4
db-dtypes==1.2.0
